% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{addScaffoldPTMFAnnotations}
\alias{addScaffoldPTMFAnnotations}
\title{Add scaffold ptm annotaitons to tmt experiment}
\usage{
addScaffoldPTMFAnnotations(eset, file)
}
\arguments{
\item{eset}{ExpressionSet}

\item{file}{path to Scaffold file}
}
\value{
ExpressionSet object
}
\description{
Add scaffold ptm annotaitons to tmt experiment
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
No references
}

