% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Targeted.R
\name{calibrationCurve}
\alias{calibrationCurve}
\title{S3 class object describing a calibration curve and storing some figures of merit}
\usage{
calibrationCurve(eset, method = "blank")
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{to calculate Limit of Detection / Limit of Quantification. c("blank","low")}
}
\value{
calibrationCurve object
}
\description{
S3 class object describing a calibration curve and storing some figures of merit
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
Statistical characterization of multiple-reaction monitoring mass spectrometry (MRM-MS) assays for quantitative proteomics, Mani et al. (2012), \url{http://www.ncbi.nlm.nih.gov/pubmed/23176545}
}

