% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphics.R
\docType{data}
\name{COLORS}
\alias{COLORS}
\title{color vector}
\format{\preformatted{ chr [1:668] "red" "darkgreen" "blue" "darkmagenta" "darkorange" ...
}}
\usage{
COLORS
}
\description{
color vector
}
\keyword{datasets}

