% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{createExpressionDataset}
\alias{createExpressionDataset}
\title{Create ExpressionSet object}
\usage{
createExpressionDataset(expressionMatrix = expressionMatrix,
  expDesign = expDesign, featureAnnotations = featureAnnotations)
}
\arguments{
\item{expressionMatrix}{matrix of expression signals per feature and sample}

\item{expDesign}{experimental design data.frame}

\item{featureAnnotations}{data.frame including e.g: Protein Description, Id score etc.}
}
\value{
ExpressionSet object
}
\description{
Create ExpressionSet object
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}

