% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{cvBoxplot}
\alias{cvBoxplot}
\title{C.V. boxplot}
\usage{
cvBoxplot(eset,
  col = as.character(.getConditionColors(eset)[unique(pData(eset)$condition),
  ]), cex.names = 0.9, cex.axis = 1.25, cex.lab = 1.25, ...)
}
\arguments{
\item{eset}{ExpressionSet}
}
\description{
C.V. boxplot
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

