% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{cysteinFreqBarplot}
\alias{cysteinFreqBarplot}
\title{Plot Cystein Frequency}
\usage{
cysteinFreqBarplot(peptides, ...)
}
\arguments{
\item{peptides}{vector}

\item{...}{see plot}
}
\description{
Plot Cystein Frequency
}
\details{
Selecting for peptides of length 7-19
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
