% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserOptions.R
\name{expDesignTagToExpDesign}
\alias{expDesignTagToExpDesign}
\title{Create experimental design data.frame from user input string}
\usage{
expDesignTagToExpDesign(tag, expDesignDefault)
}
\arguments{
\item{tag}{tag}

\item{expDesignDefault}{data.frame}
}
\value{
data.frame describing experimental design
}
\description{
Create experimental design data.frame from user input string
}
\details{
tag: 1,2:3:4,5,6
		condition isControl
	1 Condition 1 TRUE
	2 Condition 1 TRUE
	3 Condition 1 TRUE
	4 Condition 2 FALSE
	5 Condition 2 FALSE
	6 Condition 2 FALSE
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

