% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SafeQuantAnalysis.R
\name{export}
\alias{export}
\title{Export content of safeQuantAnalysis object}
\usage{
export(sqa, nbRows = nrow(sqa$pValue), file = NA)
}
\arguments{
\item{sqa}{safeQuantAnalysis object}

\item{nbRows}{Number of rows to export. Features are ordred by increasing minimal p.value}

\item{file}{file path}
}
\description{
Export content of safeQuantAnalysis object
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link{safeQuantAnalysis}}
}

