% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SafeQuantAnalysis.R
\name{export.safeQuantAnalysis}
\alias{export.safeQuantAnalysis}
\title{Export content of safeQuantAnalysis object}
\usage{
\method{export}{safeQuantAnalysis}(sqa, nbRows = nrow(sqa$pValue),
  file = NA)
}
\arguments{
\item{sqa}{safeQuantAnalysis object}

\item{nbRows}{Number of rows to export. Features are ordred by increasing minimal p.value}
}
\description{
Export content of safeQuantAnalysis object
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link{safeQuantAnalysis}}
}

