% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getAAProteinCoordinates}
\alias{getAAProteinCoordinates}
\title{Get amino acid coordinates on protein}
\usage{
getAAProteinCoordinates(peptideSeq, proteinSeq, aaRegExpr = "[STY]")
}
\arguments{
\item{peptideSeq}{peptide sequence}

\item{proteinSeq}{protein sequence}
}
\value{
vector of protein coordinates (mmodification residue number)
}
\description{
Get amino acid coordinates on protein
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

