% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getAccessionNumber}
\alias{getAccessionNumber}
\title{Extract accession numbers from Uniprot proteinNames}
\usage{
getAccessionNumber(proteinName)
}
\arguments{
\item{proteinName}{vector of protein names}
}
\value{
vector of uniprot accession numbers
}
\description{
Extract accession numbers from Uniprot proteinNames
}
\details{
sp|A0MZ66|SHOT1_HUMAN -> A0MZ66
}
\note{
No note
}
\examples{
print("No examples")
}
