% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCV}
\alias{getCV}
\title{Calculate Coefficiant of Variance per feature (Relative standard Deviation)}
\usage{
getCV(data)
}
\arguments{
\item{data}{data.frame of replicate signals}
}
\value{
vector of CVs
}
\description{
Calculate Coefficiant of Variance per feature (Relative standard Deviation)
}
\details{
CV = sd / mean
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

