% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getExpDesignProgenesisCsv}
\alias{getExpDesignProgenesisCsv}
\title{Parse Experimental Design from Progenesis Csv Export}
\usage{
getExpDesignProgenesisCsv(file)
}
\arguments{
\item{file}{path to progenesis csv file}
}
\value{
data.frame describing experimental design
}
\description{
Parse Experimental Design from Progenesis Csv Export
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

