% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getFTestPValue}
\alias{getFTestPValue}
\title{Perform statistical test (mderated F-test)}
\usage{
getFTestPValue(eset, adjust = F, log = T)
}
\arguments{
\item{eset}{ExpressionSet}

\item{adjust}{TRUE/FALSE adjust for multiple testing using Benjamini & Hochberg  (1995) method}

\item{log}{T/F log-transform expression values}
}
\value{
list of pvalues
}
\description{
Perform statistical test (mderated F-test)
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
Empirical Bayes method, Smyth (2004), \url{http://www.ncbi.nlm.nih.gov/pubmed/16646809}
}
\seealso{
\code{\link[limma]{eBayes}}
}
