% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getGeneName}
\alias{getGeneName}
\title{Extract Gene Name from uniprot fasta header description}
\usage{
getGeneName(proteinDescription)
}
\arguments{
\item{proteinDescription}{vector of descriptions}
}
\value{
vector of gene names
}
\description{
Extract Gene Name from uniprot fasta header description
}
\details{
ATP synthase subunit beta OS=Salmonella typhimurium (strain SL1344) GN=atpD -> atpD
}
\note{
No note
}
\examples{
print("No examples")
}
