% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGlobalNormFactors}
\alias{getGlobalNormFactors}
\title{Get normalization factors. calculated as median signal per run (column) over median of first run.}
\usage{
getGlobalNormFactors(eset, method = "median")
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
vector of normalization factors
}
\description{
Get normalization factors. calculated as median signal per run (column) over median of first run.
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\keyword{normalization}

