% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getGoTermDF}
\alias{getGoTermDF}
\title{Get Go term data for a list of uniprot accession numbers}
\usage{
getGoTermDF(taxId = taxId, acs = acs)
}
\arguments{
\item{taxId}{uniprot taxon identifier}

\item{acs}{vector uniprot accession numbers}
}
\value{
data.frame "UNIPROTKB" "GO-ID"     "ccIds"     "mfIds"     "bpIds"     "ccTerms"   "mfTerms"   "bpTerms"
}
\description{
Get Go term data for a list of uniprot accession numbers
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
