% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getIdLevelQvals}
\alias{getIdLevelQvals}
\title{Calculates identification level q-values based on target-decoy score distributions}
\usage{
getIdLevelQvals(scores, isDecoy)
}
\arguments{
\item{scores}{peptide/protein identficationscore}

\item{isDecoy}{vector of TRUE/FALSE}
}
\value{
vector of q.values
}
\description{
Calculates identification level q-values based on target-decoy score distributions
}
\details{
q-value = (Nb. Decoy Entries at idScore Threshold S*) / (Nb. Target Entries at idScore Threshold S). (* idScore >= S)
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

