% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TMT.R
\name{getImpuritiesMatrix}
\alias{getImpuritiesMatrix}
\title{Get Thermo TMT impurity matrix}
\usage{
getImpuritiesMatrix(plexNb = 6)
}
\arguments{
\item{plexNb}{integer, 6 or 10 plex}
}
\value{
impurity matrix matrix
}
\description{
Get Thermo TMT impurity matrix
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

