% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getImputationMatrix}
\alias{getImputationMatrix}
\title{Get matrix of imputed valeus in ExpressionSet matrix}
\usage{
getImputationMatrix(eset, method = "intensity")
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{c("intensity","count") default intensity}
}
\value{
matrix
}
\description{
Get matrix of imputed valeus in ExpressionSet matrix
}
\note{
No note
}
\examples{
print("No examples")
}
\seealso{
No note
}
