% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getImputedIntensities}
\alias{getImputedIntensities}
\title{Get matrix of imputed valeus in ExpressionSet matrix}
\usage{
getImputedIntensities(eset)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
matrix
}
\description{
Get matrix of imputed valeus in ExpressionSet matrix
}
\note{
No note
}
\examples{
print("No examples")
}
\seealso{
No note
}
