% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getKinaseFreq}
\alias{getKinaseFreq}
\title{Get kinase matching frequnecy of each phospho peptide subsequnece}
\usage{
getKinaseFreq(phosphoSeqs)
}
\arguments{
\item{phosphoSeqs}{vector of phospho peptide sub sequneces 'PARVVRpSRREEEE'}
}
\value{
ExpressionSet object
}
\description{
Get kinase matching frequnecy of each phospho peptide subsequnece
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
