% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getKinases}
\alias{getKinases}
\title{Get all kinases matching phospho peptide sub sequnece}
\usage{
getKinases(phosphoSeq)
}
\arguments{
\item{phosphoSeq}{scalar  peptide sub sequnece 'PARVVRpSRREEEE'}
}
\value{
ExpressionSet object
}
\description{
Get all kinases matching phospho peptide sub sequnece
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
