% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getLoocvFoldError}
\alias{getLoocvFoldError}
\title{Leave-One-Out Cross Validate Qunatification Model}
\usage{
getLoocvFoldError(df)
}
\arguments{
\item{df}{data.frame  of two columns 1) "signal" - ms metric 2) "cpc" absolute quantity}
}
\value{
data.frame of fold errors per (left-out) protein
}
\description{
Leave-One-Out Cross Validate Qunatification Model
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
NA
}
\keyword{normalization}

