% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getMaxIndex}
\alias{getMaxIndex}
\title{get index of max in vecotr of numeric values}
\usage{
getMaxIndex(v)
}
\arguments{
\item{v}{vector}
}
\description{
get index of max in vecotr of numeric values
}

