% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getModifProteinCoordinates}
\alias{getModifProteinCoordinates}
\title{Get modification coordinates on protein}
\usage{
getModifProteinCoordinates(modifAnnot, peptideSeq, proteinSeq)
}
\arguments{
\item{modifAnnot}{modifcation as annotated by progenesis. E.g. '[15] Phospho (ST)|[30] Phospho (ST)'}

\item{peptideSeq}{peptide sequence}

\item{proteinSeq}{protein sequence}
}
\value{
vector of protein coordinates (mmodification residue number)
}
\description{
Get modification coordinates on protein
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

