% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getMotifFreq}
\alias{getMotifFreq}
\title{Get motif matching frequnecy of each phospho peptide subsequnece}
\usage{
getMotifFreq(phosphoSeqs)
}
\arguments{
\item{phosphoSeqs}{vector of phospho peptide sub sequneces 'PARVVRpSRREEEE'}
}
\value{
ExpressionSet object
}
\description{
Get motif matching frequnecy of each phospho peptide subsequnece
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
