% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMotifX}
\alias{getMotifX}
\title{Create motif-x peptide annotation}
\usage{
getMotifX(modifAnnot, peptide, proteinSeq, motifLength = 4)
}
\arguments{
\item{peptide}{peptide sequence}

\item{proteinSeq}{protein sequence}

\item{motifLength}{motif flanking sequence}
}
\value{
vector of motifs
}
\description{
Create motif-x peptide annotation
}
\details{
motif-x example PGDYS*TTPG
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

