% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getNAFraction}
\alias{getNAFraction}
\title{Get fraction missing values per ratio/condition/run}
\usage{
getNAFraction(eset, method = "ratio")
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{c("ratio","cond","run")}
}
\value{
matrix
}
\description{
Get fraction missing values per ratio/condition/run
}
\note{
No note
}
\examples{
print("No examples")
}
\seealso{
No note
}
