% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNbPeptidesPerProtein}
\alias{getNbPeptidesPerProtein}
\title{Get number of peptides per protein}
\usage{
getNbPeptidesPerProtein(eset)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
table
}
\description{
Get number of peptides per protein
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

