% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getNbRolledUpFeatures}
\alias{getNbRolledUpFeatures}
\title{Get number rolled up features per row}
\usage{
getNbRolledUpFeatures(eset, method = "vector")
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{c("vector","matrix") default vector}
}
\value{
matrix
}
\description{
Get number rolled up features per row
}
\note{
No note
}
\examples{
print("No examples")
}
\seealso{
No note
}
