% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getNbSpectraPerProtein}
\alias{getNbSpectraPerProtein}
\title{Get number of spectra per protein}
\usage{
getNbSpectraPerProtein(eset)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
table
}
\description{
Get number of spectra per protein
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

