% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPeptides}
\alias{getPeptides}
\title{Digest protein}
\usage{
getPeptides(proteinSeq, proteaseRegExp = .getProteaseRegExp("trypsin"),
  nbMiscleavages = 0)
}
\arguments{
\item{proteinSeq}{protein sequence}
}
\value{
vector of peptides
}
\description{
Digest protein
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}

