% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TMT.R
\name{getRatioCorrectionFactorModel}
\alias{getRatioCorrectionFactorModel}
\title{Get linear model explaning log2 ratio as a function of log2 tmt ratio}
\usage{
getRatioCorrectionFactorModel(eset)
}
\arguments{
\item{eset}{paired calibration mix}
}
\value{
linear model
}
\description{
Get linear model explaning log2 ratio as a function of log2 tmt ratio
}
\details{
Uses linear model of log tmt ratio vs  log ref ratio
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

