% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getRatios}
\alias{getRatios}
\title{Calculate ratios, comparing all case to control}
\usage{
getRatios(eset, method = "median", log2 = T)
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{median or mean}

\item{log2}{transform}
}
\value{
ExpressionSet object
}
\description{
Calculate ratios, comparing all case to control
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

