% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getSignalPerCondition}
\alias{getSignalPerCondition}
\title{Summarize replicate signal per condition (min)}
\usage{
getSignalPerCondition(eset, method = "median")
}
\arguments{
\item{method}{median (default), mean, max, min, sd}

\item{data}{data.frame of replicate signals}
}
\value{
data.frame of per condition signals
}
\description{
Summarize replicate signal per condition (min)
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

