% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIA.R
\name{getSwaths}
\alias{getSwaths}
\title{get swath sizes ensuring equal number of precurosrs per swath}
\usage{
getSwaths(mzs, nbSwaths = 30, lowerOverlap = 2)
}
\arguments{
\item{mzs}{precucrsor mz values}

\item{nbSwaths}{default 30}

\item{lowerOverlap}{default 2 lower bound overlap preceeding window}
}
\value{
data.frame "binMean","lower","upper","delta"
}
\description{
get swath sizes ensuring equal number of precurosrs per swath
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
No ref
}
