% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Targeted.R
\name{ggDilutionCurve}
\alias{ggDilutionCurve}
\title{Plot dilution curve}
\usage{
ggDilutionCurve(dCurve, lod, title = "")
}
\arguments{
\item{dCurve}{data.frame columns concentration, intensity}

\item{lod}{limit of detection}

\item{title}{plot title}
}
\value{
ggplot2
}
\description{
Plot dilution curve
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
