% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGGraphics.R
\name{ggVolcanoPlot}
\alias{ggVolcanoPlot}
\title{Plots volcano, data points colored by max cv of the 2 compared conditions}
\usage{
ggVolcanoPlot(data = data, title = "", pValueThrs = 0.05,
  log2RatioThrs = 0.5849625, thrsLineCol = "lightgrey", thrsLineLty = 2,
  xlab = "log2 ratio", ylab = "-log10 pValue", textSize = 20,
  xlim = range(data$ratio, na.rm = T), ylim = range(-log10(data$pValue),
  na.rm = T), abline = c("both"), topNlabels = 10)
}
\arguments{
\item{data}{data.frame}

\item{title}{default no title}

\item{pValueThrs}{default 0.01}

\item{log2RatioThrs}{default log2(0.5)}

\item{thrsLineCol}{default "lightgrey"}

\item{xlab}{default "log2 ratio"}

\item{ylab}{default "-log10 pValue"}

\item{textSize}{default 20}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{abline}{c("none","both","ratio","pvalue")}

\item{defalut}{2}
}
\value{
ggplot2 object
}
\description{
Plots volcano, data points colored by max cv of the 2 compared conditions
}
\details{
data.frame input object should contain columns ("ratio","pValue","geneName","ac","cv", "description")
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
