% Generated by roxygen2 (4.0.1): do not edit by hand
\name{globalNormalize}
\alias{globalNormalize}
\title{Normalize, Norm factors calculated as median signal per run (column) over median of first run.}
\usage{
globalNormalize(eset, globalNormFactors)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
eset ExpressionSet
}
\description{
Normalize, Norm factors calculated as median signal per run (column) over median of first run.
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
getGlobalNormFactors
}
\keyword{normalization}

