% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hClustHeatMap}
\alias{hClustHeatMap}
\title{Hierc. clustering heat map, cluster by mms signal, display log2 ratios to control median}
\usage{
hClustHeatMap(eset, conditionColors = .getConditionColors(eset),
  selIndices = 1:nrow(eset), breaks = seq(-2, 2, length = 20), ...)
}
\arguments{
\item{eset}{ExpressionSet}

\item{conditionColors}{data.frame of colors per condition}

\item{selIndices}{indices of selected}
}
\value{
heatmap.2 obj
}
\description{
Hierc. clustering heat map, cluster by mms signal, display log2 ratios to control median
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

