% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isCon}
\alias{isCon}
\title{Check if protein is a contaminant entry}
\usage{
isCon(ac)
}
\arguments{
\item{ac}{vector of protein accession numbers}
}
\value{
vector TRUE/FALSE
}
\description{
Check if protein is a contaminant entry
}
\details{
contanminants proteins are typically annotated as: CON_P0000
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

