% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{isDecoy}
\alias{isDecoy}
\title{Check if protein is a decoy entry}
\usage{
isDecoy(ac)
}
\arguments{
\item{ac}{vector of protein accession numbers}
}
\value{
vector TRUE/FALSE
}
\description{
Check if protein is a decoy entry
}
\details{
decoy proteins are typically annotated as: REV_P0000
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

