% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserOptions.R
\docType{data}
\name{option_list}
\alias{option_list}
\title{Command Line Option List}
\format{\preformatted{List of 25
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr "-i"
  .. ..@ long_flag : chr "--inputFile"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "inputFile"
  .. ..@ default   : chr ""
  .. ..@ help      : chr "I/O:  Input file: Progenesis (Feature,Protein or Peptide) .csv,\\n\\t\\t\\tor Scaffold Q+ (Raw Export, for TMT quant) .xls (REQUIRE"| __truncated__
  .. ..@ metavar   : chr "inputFile"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr "-o"
  .. ..@ long_flag : chr "--outputDir"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "outputDir"
  .. ..@ default   : chr NA
  .. ..@ help      : chr "I/O:  Results Output Directory [default FOLDER OF INPUTFILE]"
  .. ..@ metavar   : chr "outputDir"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr "-l"
  .. ..@ long_flag : chr "--resultsFileLabel"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "resultsFileLabel"
  .. ..@ default   : chr "SQ_Results"
  .. ..@ help      : chr "I/O: results file directory [default \%default]"
  .. ..@ metavar   : chr "resultsFileLabel"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr "-f"
  .. ..@ long_flag : chr "--fastaFile"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "fastaFile"
  .. ..@ default   : chr ""
  .. ..@ help      : chr "I/O:  Protein DB .fasta file [default ./]"
  .. ..@ metavar   : chr "fastaFile"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr "-p"
  .. ..@ long_flag : chr "--scaffoldPTMSpectrumReportFile"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "scaffoldPTMSpectrumReportFile"
  .. ..@ default   : chr ""
  .. ..@ help      : chr "I/O:  Scaffold PTM Spectrum Report File [default ./]"
  .. ..@ metavar   : chr "scaffoldPTMSpectrumReportFile"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FProteinAccessionSelection"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "FProteinAccessionSelection"
  .. ..@ default   : chr "."
  .. ..@ help      : chr "FILTER: --FP Filter features by Accession Regular Expression [default \%default] (all features kept)"
  .. ..@ metavar   : chr "Protein Accession Reg. expr."
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FModificationSelection"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "FModificationSelection"
  .. ..@ default   : chr ""
  .. ..@ help      : chr "FILTER (LFQ PEP ONLY): --FM Only keep Peptides with modifications matching Regular Expression [default \%default]\\n\\t\\t\\t\\t (all"| __truncated__
  .. ..@ metavar   : chr "modification name Reg. expr."
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FFdrCutoff"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "double"
  .. ..@ dest      : chr "FFdrCutoff"
  .. ..@ default   : num 0.01
  .. ..@ help      : chr "FILTER (LFQ ONLY): --FF Identification level False Discovery Rate Cutoff.  [0-1] [default \%default]"
  .. ..@ metavar   : chr "Peptide/Protein FDR cutoff"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FCoefficientOfVarianceMax"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "double"
  .. ..@ dest      : chr "FCoefficientOfVarianceMax"
  .. ..@ default   : num Inf
  .. ..@ help      : chr "FILTER: --FC Do not include features with C.V. above this threshold in statistical \\n\\t\\t\\t\\ttest for differential expression ["| __truncated__
  .. ..@ metavar   : chr "Coefficent of Variance cutoff"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FDeltaMassTolerancePrecursor"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "FDeltaMassTolerancePrecursor"
  .. ..@ default   : chr "AUTO SET"
  .. ..@ help      : chr "FILTER (LFQ PEP ONLY): --FD Precursor mass Error Range filter (ppm) [default \%default].\\n\\t\\t\\t\\tPeptide imports ONLY"
  .. ..@ metavar   : chr "Mass Range [x,y]"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FNumberOfPeptidesPerProteinMin"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "integer"
  .. ..@ dest      : chr "FNumberOfPeptidesPerProteinMin"
  .. ..@ default   : int 1
  .. ..@ help      : chr "FILTER: --FN Only include those proteins with at least x identified peptides [default \%default]\\n\\t\\t\\t\\tProtein analysis ONLY."| __truncated__
  .. ..@ metavar   : chr "Number of peptides"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FSitesPerPeptide"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "integer"
  .. ..@ dest      : chr "FSitesPerPeptide"
  .. ..@ default   : int 99999
  .. ..@ help      : chr "FILTER: --FS Max Nb. Modifications Per Peptide [default Inf]\\n\\t\\t\\t\\t\\t\\tPeptide analysis ONLY."
  .. ..@ metavar   : chr "Max Number of PTM sites Per Petptide"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--FLengthPeptide"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "integer"
  .. ..@ dest      : chr "FLengthPeptide"
  .. ..@ default   : int 1
  .. ..@ help      : chr "FILTER: --FL Min Peptide Length (Nb. AA's) [default Inf]\\n\\t\\t\\t\\t\\t\\tPeptide analysis ONLY."
  .. ..@ metavar   : chr "Min Peptide Length (>=)"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--SAnchorProtein"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "SAnchorProtein"
  .. ..@ default   : chr "."
  .. ..@ help      : chr "STATISTICS: --SA Normalize Intensities by selected protein(s) Regular Expression\\n\\t\\t\\t [default \%default] (use all proteins)."| __truncated__
  .. ..@ metavar   : chr "Protein Accession Reg. expr."
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--SPvalueInclude"
  .. ..@ action    : chr "store_true"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "SPvalueInclude"
  .. ..@ default   : logi FALSE
  .. ..@ help      : chr "STATISTICS: --SP output eBayes moderated t-statistic p-values [default \%default]"
  .. ..@ metavar   : chr(0) 
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--SNonPairWiseStatTest"
  .. ..@ action    : chr "store_true"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "SNonPairWiseStatTest"
  .. ..@ default   : logi FALSE
  .. ..@ help      : chr "STATISTICS: --SN non pairwise eBayes moderated t-statistic p-values [default \%default]"
  .. ..@ metavar   : chr(0) 
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--EXperimentalDesign"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "character"
  .. ..@ dest      : chr "EXperimentalDesign"
  .. ..@ default   : chr NA
  .. ..@ help      : chr "EXPERIMENTAL DESIGN: --EX \\",\\" seperated samples, \\":\\" separated conditions \\n\\t\\t\\t\\t\\tExample: 1,2,3:4,5,6 \\n\\t\\t\\t\\t\\t   c"| __truncated__
  .. ..@ metavar   : chr "EXperimentalDesign"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--EProteinQuantOff"
  .. ..@ action    : chr "store_false"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "EProteinQuantOff"
  .. ..@ default   : logi TRUE
  .. ..@ help      : chr "EXPERIMENTAL DESIGN: --EP Disable Protein Level Quantification [default \%default]"
  .. ..@ metavar   : chr(0) 
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--PRatioCutOff"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "double"
  .. ..@ dest      : chr "PRatioCutOff"
  .. ..@ default   : num 1
  .. ..@ help      : chr "PDF-REPORT: --PR Intensity ratio (fold change) cut-off used for graphics export. >1 [default \%default]"
  .. ..@ metavar   : chr "Intensity ratio cutoff"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--PQvaueCutOff"
  .. ..@ action    : chr "store"
  .. ..@ type      : chr "double"
  .. ..@ dest      : chr "PQvaueCutOff"
  .. ..@ default   : num 0.01
  .. ..@ help      : chr "PDF-REPORT: --PQ Qvalue cut-off used for graphics. \\n\\t\\t\\tHigh-lighting features with a qval < specified value. [0-1] [default"| __truncated__
  .. ..@ metavar   : chr "Differential expression qvalue cutOff"
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--ARDataFile"
  .. ..@ action    : chr "store_true"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "ARDataFile"
  .. ..@ default   : logi FALSE
  .. ..@ help      : chr "ADDITIONAL-REPORTS: --AR Save R objects in 'label'.RData file [default \%default]"
  .. ..@ metavar   : chr(0) 
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--AIbaq"
  .. ..@ action    : chr "store_true"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "AIbaq"
  .. ..@ default   : logi FALSE
  .. ..@ help      : chr "ADDITIONAL-REPORTS (LFQ PROT): --AI creates .tsv output file\\n\\t\\t\\t\\t\\tincluding protein iBAQ values. [default \%default]"
  .. ..@ metavar   : chr(0) 
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr NA
  .. ..@ long_flag : chr "--ATop3"
  .. ..@ action    : chr "store_true"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "ATop3"
  .. ..@ default   : logi FALSE
  .. ..@ help      : chr "ADDITIONAL-REPORTS (LFQ PEP): --AT creates .tsv output file\\n\\t\\t\\t\\t\\tincluding protein top3 values. [default \%default]"
  .. ..@ metavar   : chr(0) 
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr "-t"
  .. ..@ long_flag : chr "--test"
  .. ..@ action    : chr "store_true"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "test"
  .. ..@ default   : logi FALSE
  .. ..@ help      : chr "TEST: test option, include first 2000 entries only [default \%default]\\n\\t\\t\\tPeptide analysis ONLY."
  .. ..@ metavar   : chr(0) 
 $ :Formal class 'OptionParserOption' [package "optparse"] with 8 slots
  .. ..@ short_flag: chr "-v"
  .. ..@ long_flag : chr "--verbose"
  .. ..@ action    : chr "store_true"
  .. ..@ type      : chr "logical"
  .. ..@ dest      : chr "verbose"
  .. ..@ default   : logi FALSE
  .. ..@ help      : chr "Print extra output [default \%default]"
  .. ..@ metavar   : chr(0) 
}}
\usage{
option_list
}
\description{
Command Line Option List
}
\keyword{datasets}

