% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseMaxQuantProteinGroupTxt}
\alias{parseMaxQuantProteinGroupTxt}
\title{Parse MaxQuant Protein Group Txt}
\usage{
parseMaxQuantProteinGroupTxt(file = file, expDesign = expDesign,
  method = "auc")
}
\arguments{
\item{file}{path to MaxQuant Protein txt file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}
}
\value{
ExpressionSet object
}
\description{
Parse MaxQuant Protein Group Txt
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}

