% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parseProgenesisFeatureCsv}
\alias{parseProgenesisFeatureCsv}
\title{Parse Progenesis Peptide Csv}
\usage{
parseProgenesisFeatureCsv(file = file,
  expDesign = getExpDesignProgenesisCsv(file), method = "auc")
}
\arguments{
\item{file}{path to Progenesis Peptide csv file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}
}
\value{
ExpressionSet object
}
\description{
Parse Progenesis Peptide Csv
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}

