% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parseProgenesisPeptideCsv}
\alias{parseProgenesisPeptideCsv}
\title{Parse Progenesis Peptide Csv}
\usage{
parseProgenesisPeptideCsv(file = file, expDesign = NA)
}
\arguments{
\item{file}{path to Progenesis Peptide csv file}

\item{expDesign}{experimental design data.frame}
}
\value{
ExpressionSet object
}
\description{
Parse Progenesis Peptide Csv
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}

