% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseProgenesisPeptideMeasurementCsv}
\alias{parseProgenesisPeptideMeasurementCsv}
\title{Parse Progenesis Peptide Measurement Csv Export}
\usage{
parseProgenesisPeptideMeasurementCsv(file, expDesign = expDesign,
  method = "auc",
  expressionColIndices = .getProgenesisCsvExpressionColIndices(file, method =
  method))
}
\arguments{
\item{file}{path to Progenesis Peptide Measurement csv file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}

\item{expressionColIndices}{default .getProgenesisCsvExpressionColIndices()}
}
\value{
ExpressionSet object
}
\description{
Parse Progenesis Peptide Measurement Csv Export
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}

