% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseScaffoldPTMReport}
\alias{parseScaffoldPTMReport}
\title{Parse scaffold PTM Spectrum Report}
\usage{
parseScaffoldPTMReport(file)
}
\arguments{
\item{file}{path to Scaffold file}
}
\value{
data.frame
}
\description{
Parse scaffold PTM Spectrum Report
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

