% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGGraphics.R
\name{plotAllGGVolcanoes}
\alias{plotAllGGVolcanoes}
\title{Plots volcano of all condition comparisons}
\usage{
plotAllGGVolcanoes(sqa, isAdjusted = T, ...)
}
\arguments{
\item{sqa}{SafeQuantAnalysis object}

\item{isAdjusted}{(T/F) plot adjusted pvalues}

\item{see}{ggVolcanoPlot}
}
\value{
ggplot2 object
}
\description{
Plots volcano of all condition comparisons
}
\details{
data.frame input object should contain columns ("ratio","pValue","geneName","ac","cv", "description")
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
