% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCalibrationCurve}
\alias{plotCalibrationCurve}
\title{Plot calibration Curve}
\usage{
plotCalibrationCurve(fit, dispElements = c("formula", "lowess", "stats"),
  xlab = "Conc. (CPC) ", ylab = "Pred. Conc.",
  predictorName = paste("log10(", names(coef(fit))[2], ")", sep = ""),
  text = F, ..., main = "")
}
\arguments{
\item{fit}{simple log-linear model}

\item{dispElements}{c("formula","lowess","stats")}
}
\description{
Plot calibration Curve
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

