% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotExpDesign}
\alias{plotExpDesign}
\title{Display experimental design, high-lighting the control condition}
\usage{
plotExpDesign(eset, condColors = .getConditionColors(eset), version = "X")
}
\arguments{
\item{eset}{ExpressionSet}
}
\description{
Display experimental design, high-lighting the control condition
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

