% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotIdScoreVsFDR}
\alias{plotIdScoreVsFDR}
\title{Plot FDR vs. identification score}
\usage{
plotIdScoreVsFDR(idScore, qvals, ylab = "False Discovery Rate",
  xlab = "Identification Score", ...)
}
\arguments{
\item{idScore}{vector of identification scores}

\item{qvals}{vector of q-valres}
}
\description{
Plot FDR vs. identification score
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

