% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotNbIdentificationsVsRT}
\alias{plotNbIdentificationsVsRT}
\title{Plot the number of identified Features per Reteintion Time minute.}
\usage{
plotNbIdentificationsVsRT(eset, cex.axis = 1.25, cex.lab = 1.25,
  col = "blue", lwd = 2, ...)
}
\arguments{
\item{eset}{ExpressionSet}
}
\description{
Plot the number of identified Features per Reteintion Time minute.
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

