% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPrecMassErrorDistrib}
\alias{plotPrecMassErrorDistrib}
\title{Plot Precursor Mass Error Distribution}
\usage{
plotPrecMassErrorDistrib(pMassError, isDecoy, pMassTolWindow = c(-10, 10))
}
\arguments{
\item{pMassError}{vector of precursor mass errors}

\item{isDecoy}{vector TRUE/FALSE}

\item{pMassTolWindow}{Precursor Mass Error Tolerance Window}
}
\description{
Plot Precursor Mass Error Distribution
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

