% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPrecMassErrorVsScore}
\alias{plotPrecMassErrorVsScore}
\title{Plot precursorMass error v.s score highlighting decoy and displaying user specified user specified precursor mass filter}
\usage{
plotPrecMassErrorVsScore(pMassError, idScore, isDecoy, pMassTolWindow = c(-10,
  10))
}
\arguments{
\item{pMassError}{vector of precursor mass errors}

\item{idScore}{vector of identification scores}

\item{isDecoy}{vector TRUE/FALSE}

\item{pMassTolWindow}{Precursor Mass Error Tolerance Window}
}
\description{
Plot precursorMass error v.s score highlighting decoy and displaying user specified user specified precursor mass filter
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

