% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotQValueVsPValue}
\alias{plotQValueVsPValue}
\title{Plot qValue vs pValue}
\usage{
plotQValueVsPValue(sqa, lim = c(0, 1), ...)
}
\arguments{
\item{sqa}{SafeQuantAnalysis Object}

\item{lim}{x-axis and y-axis range}

\item{...}{see plot}
}
\description{
Plot qValue vs pValue
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

