% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotROC}
\alias{plotROC}
\title{Plot Number of Identifications vs. FDR}
\usage{
plotROC(qvals, xlab = "False Discovery Rate",
  ylab = "# Valid Identifications", xlim = c(0, 1), breaks = 100,
  col = "blue", lwd = 1.5, ...)
}
\arguments{
\item{qvals}{vector of q-values}

\item{breaks}{see breaks for hist function}
}
\description{
Plot Number of Identifications vs. FDR
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

