% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotRTNorm}
\alias{plotRTNorm}
\title{Plot all retention time profile overalying ratios}
\usage{
plotRTNorm(rtNormFactors, eset, samples = 1:ncol(rtNormFactors), main = "",
  ...)
}
\arguments{
\item{rtNormFactors}{data.frame of normalization factor per r.t bin and sample, obtained by getRTNormFactors}

\item{eset}{ExprsssionSet}

\item{samples}{specify samples (sample numbers) to be plotted}
}
\description{
Plot all retention time profile overalying ratios
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
In Silico Instrumental Response Correction Improves Precision of Label-free Proteomics and Accuracy of Proteomics-based Predictive Models, Lyutvinskiy et al. (2013), \url{http://www.ncbi.nlm.nih.gov/pubmed/23589346}
}
\seealso{
\code{\link{getRTNormFactors}}
}

