% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotRTNormSummary}
\alias{plotRTNormSummary}
\title{Plot all retention time normalization profiles}
\usage{
plotRTNormSummary(eset,
  col = as.character(.getConditionColors(eset)[pData(eset)$condition, 1]),
  ...)
}
\arguments{
\item{rtNormFactors}{data.frame of normalization factor per r.t bin and sample, obtained by getRTNormFactors}

\item{condNames}{vector of condition names}
}
\description{
Plot all retention time normalization profiles
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
In Silico Instrumental Response Correction Improves Precision of Label-free Proteomics and Accuracy of Proteomics-based Predictive Models, Lyutvinskiy et al. (2013), \url{http://www.ncbi.nlm.nih.gov/pubmed/23589346}
}
\seealso{
\code{\link{getRTNormFactors}}
}

