% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotVolcano}
\alias{plotVolcano}
\title{Plots volcano, data points colored by max cv of the 2 compared conditions}
\usage{
plotVolcano(obj, ratioCutOffAbsLog2 = 0, absLog10pValueCutOff = 2,
  adjusted = T, ...)
}
\arguments{
\item{obj}{safeQuantAnalysis object or data.frame}

\item{ratioCutOffAbsLog2}{ratio abline}

\item{absLog10pValueCutOff}{pValue abline}

\item{adjusted}{TRUE/FALSE plot qValues or pValues on y-axis}
}
\description{
Plots volcano, data points colored by max cv of the 2 compared conditions
}
\details{
data.frame input object should contain 3 columns (ratio,qValue,cv)
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

