% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rollUp}
\alias{rollUp}
\title{Roll up feature intensites per unique colum combination}
\usage{
rollUp(eset = eset, featureDataColumnName = c("peptide"),
  method = c("sum"), isProgressBar = F)
}
\arguments{
\item{eset}{ExpressionSet}

\item{featureDataColumnName}{vector of column names e.g. peptide or proteinName}

\item{method}{"sum", "mean" or "top3"}

\item{isProgressBar}{TRUE/FALSE display progress bar}
}
\value{
ExpressionSet object
}
\description{
Roll up feature intensites per unique colum combination
}
\details{
featureDataColumnName = c("peptide","charge","ptm"), method= c("sum"), sums up intensities per peptie modification charge state
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
No references
}
\seealso{
\code{\link{topX}}
}

