% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{rtNormalize}
\alias{rtNormalize}
\title{Normalization data per retention time bin}
\usage{
rtNormalize(eset, rtNormFactors)
}
\arguments{
\item{eset}{ExpressionSet}

\item{rtNormFactors}{obtained using getRTNormFactors}
}
\value{
data.frame normalization factors per retention time bin (minute)
}
\description{
Normalization data per retention time bin
}
\details{
Normalize for variations in elelctrospray ionization current.
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
In Silico Instrumental Response Correction Improves Precision of Label-free Proteomics and Accuracy of Proteomics-based Predictive Models, Lyutvinskiy et al. (2013), \url{http://www.ncbi.nlm.nih.gov/pubmed/23589346}
}
\seealso{
\code{\link{getRTNormFactors}}
}

