% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setNbPeptidesPerProtein}
\alias{setNbPeptidesPerProtein}
\title{Set nbPeptides coulmn of featureData}
\usage{
setNbPeptidesPerProtein(eset)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
eset
}
\description{
Set nbPeptides coulmn of featureData
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

