% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{setNbSpectraPerProtein}
\alias{setNbSpectraPerProtein}
\title{Set nbPeptides coulmn of featureData}
\usage{
setNbSpectraPerProtein(eset)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
eset
}
\description{
Set nbPeptides coulmn of featureData
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

