% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sqNormalize}
\alias{sqNormalize}
\title{Normalize}
\usage{
sqNormalize(eset, method = "global")
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{c("global","rt")}
}
\value{
eset ExpressionSet
}
\description{
Normalize
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
getGlobalNormFactors, getRTNormFactors
}
\keyword{normalization}

