% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{standardise}
\alias{standardise}
\title{Standardise data}
\usage{
standardise(d)
}
\arguments{
\item{d}{vector or data.frame or matrix}
}
\value{
vector or data.frame or matrix
}
\description{
Standardise data
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}

