% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{barplotMSSignal}
\alias{barplotMSSignal}
\title{Barplot of ms-signal per column}
\usage{
barplotMSSignal(eset,
  col = as.character(.getConditionColors(eset)[pData(eset)$condition, ]),
  method = c("sum", "sharedSignal"), cex.lab = 1.25, cex.axis = 1.25,
  labels = rownames(pData(eset)), ...)
}
\arguments{
\item{method}{c("median","sum","sharedSignal")}

\item{matrix}{matrix of ms-signals}

\item{color}{color}
}
\description{
Barplot of ms-signal per column
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

