% Generated by roxygen2 (4.0.1): do not edit by hand
\name{barplotMSSignal}
\alias{barplotMSSignal}
\title{Barplot of ms-signal sums per column}
\usage{
barplotMSSignal(eset, cex.lab = 1, ...)
}
\arguments{
\item{matrix}{matrix of ms-signals}

\item{color}{color}
}
\description{
Barplot of ms-signal sums per column
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

