% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TMT.R
\name{createExpDesign}
\alias{createExpDesign}
\title{Create Experimental Design}
\usage{
createExpDesign(tag, nbPlex)
}
\arguments{
\item{tag}{user input tag e.g. 1,2,3:4,5,6 indicating two condition with 3 reps each}

\item{nbPlex}{tmt 6  or 10 plex}
}
\value{
expDesign data.frame
}
\description{
Create Experimental Design
}
\details{
The first listed condition is always the control condition
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

