% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Targeted.R
\name{getAllDotProduct}
\alias{getAllDotProduct}
\title{Return dotProducts to most transition intensities of most intense runs}
\usage{
getAllDotProduct(eset, nbRefRuns = 4)
}
\arguments{
\item{eset}{ExpressionSet}

\item{nbRefRuns}{(default top  4)}

\item{u}{vector 1}

\item{v}{vector 2}

\item{dp}{matrix of dp's per peptide and run}
}
\value{
dp
}
\description{
Return dotProducts to most transition intensities of most intense runs
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

