% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getAllEBayes}
\alias{getAllEBayes}
\title{Perform statistical test (mderated t-test), comparing all case to control}
\usage{
getAllEBayes(eset = eset, adjust = F, log = T, method = "pairwise")
}
\arguments{
\item{eset}{ExpressionSet}

\item{adjust}{TRUE/FALSE adjust for multiple testing using Benjamini & Hochberg  (1995) method}

\item{log}{T/F log-transform expression values}

\item{method}{c("pairwise","all")}
}
\value{
ExpressionSet object
}
\description{
Perform statistical test (mderated t-test), comparing all case to control
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
Empirical Bayes method, Smyth (2004), \url{http://www.ncbi.nlm.nih.gov/pubmed/16646809}
}
\seealso{
\code{\link[limma]{eBayes}}
}

