% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{getExpDesignProgenesisCsv}
\alias{getExpDesignProgenesisCsv}
\title{Parse Experimental Design from Progenesis Csv Export}
\usage{
getExpDesignProgenesisCsv(file,
  expressionColIndices = .getProgenesisCsvExpressionColIndices(file))
}
\arguments{
\item{file}{path to progenesis csv file}

\item{expressionColIndices}{default .getProgenesisCsvExpressionColIndices(file)}
}
\value{
data.frame describing experimental design
}
\description{
Parse Experimental Design from Progenesis Csv Export
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

