% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{getGlobalNormFactors}
\alias{getGlobalNormFactors}
\title{Get normalization factors. calculated as summed/median signal per run (column) over summed/median of first run.}
\usage{
getGlobalNormFactors(eset, method = "median")
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{c("sum","median)}
}
\value{
vector of normalization factors
}
\description{
Get normalization factors. calculated as summed/median signal per run (column) over summed/median of first run.
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\keyword{normalization}

