% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getNbDetectablePeptides}
\alias{getNbDetectablePeptides}
\title{Get number peptides passing defined length criteria}
\usage{
getNbDetectablePeptides(peptides, peptideLength = c(5, 36))
}
\arguments{
\item{peptides}{list of peptides}

\item{peptideLength}{vector of two integers defining peptide length range}
}
\value{
integer corresponding to number of detectable peptides
}
\description{
Get number peptides passing defined length criteria
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}

