% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getNbMisCleavages}
\alias{getNbMisCleavages}
\title{Get number of mis-cleavages perp peptide}
\usage{
getNbMisCleavages(peptide, protease = "trypsin")
}
\arguments{
\item{peptide}{character vector}

\item{protease}{regular expression}
}
\value{
vector ofintegers
}
\description{
Get number of mis-cleavages perp peptide
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

