% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationAnalysis.R
\name{getNbPeptidesPerProtein}
\alias{getNbPeptidesPerProtein}
\title{Get number of peptides per protein}
\usage{
getNbPeptidesPerProtein(eset)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
table
}
\description{
Get number of peptides per protein
}
\details{
NA
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

