% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{hClustHeatMap}
\alias{hClustHeatMap}
\title{Hierarchical clustering heat map, cluster by and display log2 ratios to control median}
\usage{
hClustHeatMap(eset, conditionColors = .getConditionColors(eset),
  breaks = seq(-2, 2, length = 20), ...)
}
\arguments{
\item{eset}{ExpressionSet}

\item{conditionColors}{data.frame of colors per condition}
}
\description{
Hierarchical clustering heat map, cluster by and display log2 ratios to control median
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

