% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{pairsAnnot}
\alias{pairsAnnot}
\title{Plot lower triangle Pearsons R^2. Diagonal text, upper triangle all against all scatter plots with lm abline}
\usage{
pairsAnnot(data, textCol = rep(1, ncol(data)), diagText = c(), ...)
}
\arguments{
\item{data}{data.frame}
}
\description{
Plot lower triangle Pearsons R^2. Diagonal text, upper triangle all against all scatter plots with lm abline
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

