% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseProgenesisProteinCsv}
\alias{parseProgenesisProteinCsv}
\title{Parse Progenesis Protein Csv}
\usage{
parseProgenesisProteinCsv(file = file, expDesign = expDesign,
  method = "auc")
}
\arguments{
\item{file}{path to Progenesis Protein csv file}

\item{expDesign}{experimental design data.frame}

\item{method}{auc (area under curve) or spc (spectral count)}
}
\value{
ExpressionSet object
}
\description{
Parse Progenesis Protein Csv
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}

