% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotExpDesign}
\alias{plotExpDesign}
\title{Display experimental design, high-lighting the control condition}
\usage{
plotExpDesign(eset, condColors = .getConditionColors(eset), version = "X")
}
\arguments{
\item{eset}{ExpressionSet}

\item{condColors}{condition colors}

\item{version}{version number}
}
\description{
Display experimental design, high-lighting the control condition
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

