% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotIdScoreVsFDR}
\alias{plotIdScoreVsFDR}
\title{Plot FDR vs. identification score}
\usage{
plotIdScoreVsFDR(idScore, qvals, qvalueThrs = 0.01,
  ylab = "False Discovery Rate", xlab = "Identification Score", ...)
}
\arguments{
\item{idScore}{vector of identification scores}

\item{qvals}{vector of q-valres}

\item{qvalueThrs}{threshold indicated by horizontal line}

\item{ylab}{default False Discovery Rate}

\item{xlab}{default Identification Score}

\item{...}{see plot}
}
\description{
Plot FDR vs. identification score
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

