% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMSSignalDistributions}
\alias{plotMSSignalDistributions}
\title{Plot ms.signal distributions}
\usage{
plotMSSignalDistributions(d, col = 1:100, cex.axis = 1, cex.lab = 1,
  ylab = "Frequnecy", xlab = "MS-Signal", ...)
}
\arguments{
\item{matrix}{matrix of ms-signals}

\item{color}{color}
}
\description{
Plot ms.signal distributions
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

