% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotPrecMassErrorVsScore}
\alias{plotPrecMassErrorVsScore}
\title{Plot precursorMass error v.s score highlighting decoy and displaying user specified user specified precursor mass filter}
\usage{
plotPrecMassErrorVsScore(eset, pMassTolWindow = c(-10, 10), ...)
}
\arguments{
\item{eset}{ExpressionSet}

\item{pMassTolWindow}{Precursor Mass Error Tolerance Window}

\item{...}{see plot}
}
\description{
Plot precursorMass error v.s score highlighting decoy and displaying user specified user specified precursor mass filter
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

