% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotRTNormSummary}
\alias{plotRTNormSummary}
\title{Plot all retention time normalization profiles}
\usage{
plotRTNormSummary(rtNormFactors, ylim = range(rtNormFactors),
  col = 1:ncol(rtNormFactors), condNames = paste("Cond",
  1:length(unique(col))), ...)
}
\arguments{
\item{rtNormFactors}{data.frame of normalization factor per r.t bin and sample, obtained by getRTNormFactors}

\item{condNames}{vector of condition names}
}
\description{
Plot all retention time normalization profiles
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
In Silico Instrumental Response Correction Improves Precision of Label-free Proteomics and Accuracy of Proteomics-based Predictive Models, Lyutvinskiy et al. (2013), \url{http://www.ncbi.nlm.nih.gov/pubmed/23589346}
}
\seealso{
\code{\link{getRTNormFactors}}
}

