% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plots volcano, data points colored by max cv of the 2 compared conditions}
\usage{
plotVolcano(obj, ratioThrs = 1, pValueThreshold = 0.01, adjusted = T, ...)
}
\arguments{
\item{obj}{safeQuantAnalysis object or data.frame}

\item{ratioThrs}{default 1}

\item{pValueThreshold}{default 0.01}

\item{adjusted}{TRUE/FALSE plot qValues or pValues on y-axis}

\item{...}{see plot}
}
\description{
Plots volcano, data points colored by max cv of the 2 compared conditions
}
\details{
data.frame input object should contain 3 columns (ratio,qValue,cv)
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

