% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plotXYDensity}
\alias{plotXYDensity}
\title{Scatter plot with density coloring}
\usage{
plotXYDensity(x, y, isFitLm = T, legendPos = "bottomright",
  disp = c("abline", "R"), ...)
}
\arguments{
\item{x}{number vector}

\item{y}{number vector}
}
\description{
Scatter plot with density coloring
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}

