% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ExpressionAnalysis.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Set value to NA if it deviatves with more than 1.5 * IQR from lower/upper quantile}
\usage{
removeOutliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{vector}{numeric}

\item{a}{logical indicating whether missing values should be removed.}
}
\value{
vector numeric
}
\description{
Set value to NA if it deviatves with more than 1.5 * IQR from lower/upper quantile
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
\seealso{
NA
}
\keyword{normalization}

