% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SafeQuantAnalysis.R
\name{safeQuantAnalysis}
\alias{safeQuantAnalysis}
\title{safeQunat s3 class}
\usage{
safeQuantAnalysis(eset = eset, method = c("global", "naRep", "pairwise"))
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{c("global","naRep","pairwise")}
}
\description{
safeQunat s3 class
}

