% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeQuantAnalysis.R
\name{safeQuantAnalysis}
\alias{safeQuantAnalysis}
\title{safeQunat s3 class}
\usage{
safeQuantAnalysis(eset = eset, method = c("global", "naRep", "pairwise"),
  intensityAdjustmentObj = NA, fcThrs = 1)
}
\arguments{
\item{eset}{ExpressionSet}

\item{method}{c("global","naRep","rt","quantile","pairwise","all)}

\item{intensityAdjustmentObj}{list}

\item{fcThrs}{fold change threshold}
}
\description{
safeQunat s3 class
}

